Rasa_Chatbot(rerank)/requirements.txt
chromadb==0.6.3
deep_translator==1.11.4
fastapi==0.115.8
langchain==0.3.19
langchain_openai==0.3.7
langdetect==1.0.9
numpy==2.2.3
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
rank_bm25==0.2.2
rasa_sdk==3.11.0
Requests==2.32.3
sentence_transformers==3.4.1
spacy==3.8.3
transformers==4.47.1
uvicorn==0.34.0
